--- src/pages/Product/Detail/Detail.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Detail = () => {
-  return <h1>Detail</h1>;
-};
-
-export default Detail;
