+++ src/pages/Product/Product.scss
@@ -31,7 +31,7 @@ a {
       justify-content: space-between;
       align-items: center;
       padding-bottom: 10px;
-      border-bottom: 2px solid #ececec;
+      border-bottom: 2px solid #ddd;
       .productSortBar {
         display: flex;
         justify-content: flex-end;
